soccor-plymarket/polymarket-soccer-edge-agent/apps/api/pyproject.toml
[project]
name = "polymarket-soccer-edge-api"
version = "0.1.0"
description = "FastAPI backend for the Polymarket Soccer Edge Agent."
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.32",
  "psycopg[binary]>=3.2.1",
  "httpx>=0.27.0",
  "tenacity>=8.5.0",
  "redis>=5.0.8",
  "rq>=1.16.2",
  "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.23.8",
  "ruff>=0.6.2",
  "mypy>=1.11.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

